requirements.txt
fastapi
uvicorn
pdfplumber
python-multipart
sentence-transformers
python-dotenv
crewai
openai
spacy
faiss-cpu
langchain
httpx
numpy==1.24.4
pydantic
scikit-learn
aiofile
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
tqdm>=4.65.0
lxml>=4.9.0
urllib3>=2.0.0
streamlit>=1.28.0
plotly>=5.15.0
google-api-python-client>=2.100.0